unmatched_donors_django/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-crispy-forms==2.1
crispy-bootstrap5==2.0.2
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.15
